Update Dockerfile
Reffering to #157 I build it from this Dockerfile which git clones the latest code from github. Removed bash form the apk add command as I do not see the need for it. maybe I am wrong, but it works for me. So with this Dockerfile iI am running 1.4.0 right now without a problem.