Improve configure issues with SSIDs that contain blanks
The wifi networks returned by ConfigureAmazonDash.get_networks_availables()
are escaped and so it's necessary to escape the ssid when searching
the list.

This allowed me to *find* the network but sending the ssid as:

    /downloads/?amzn_ssid=My+Network&amzn_pw=SimplePassword

didn't work so I tried using urllib.parse.quote() to do the quoting:

    /downloads/?amzn_ssid=My%20Network&amzn_pw=SimplePassword

which doesn't work either.

I know dash buttons work with SSIDs that contain blanks because I
currently have 15 of them on such a network (and I'm trying to move
them a different network).